Improve behavior of auth domain fallback view
Improve the behavior of the auth domain fallback view to attempt
redirecting to the main domain instead of just returning a community not
found page.

This is useful as error views by the OAuth library are still rendered on
the auth domain, but they also contain a link to go back to the main
page, which would previously be a blank page. After this change, there's
a reasonable chance the user will end up at least to the front page of
entire Thunderstore.